requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
numpy==1.24.3
faiss-cpu==1.7.4
python-docx==1.1.0
pypdf==3.17.4
bcrypt==4.1.2
aiofiles==23.2.1
python-multipart==0.0.6
PyJWT==2.8.0
google-generativeai==0.3.2
